LayoutTests/fast/dom/HTMLAnchorElement/anchor-text-attribute-expected.txt

Tests for HTMLAnchorElement text attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a.textContent is ""
PASS a.text is ""
PASS a.textContent is "text1"
PASS a.text is "text1"
PASS a.textContent is "text2"
PASS a.text is "text2"
PASS a.text is "ab"
PASS a.textContent is "ab"
PASS successfullyParsed is true

TEST COMPLETE
